# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
...
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      telegram_id:
        example: 123456789
        type: integer
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Максим Максимович
        type: string
      phone_number:
        example: "+71112223344"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      telegram_id:
        example: 123456789
        type: integer
      version:
        example: 3
        type: integer
    type: object
...
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
...
  /users/{id}:
    delete:
      description: Удаление существующего пользователя по его ID
      parameters:
      - description: ID удаляемого пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Нет или неверные данные Telegram
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Только для администратора
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: У пользователя есть задачи
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - TelegramInitData: []
      summary: Удаление пользователя
      tags:
      - users
    get:
      description: Получение конкретного пользователя по его ID
      parameters:
      - description: ID получаемого пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно найден
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Нет или неверные данные Telegram
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - TelegramInitData: []
      summary: Получение пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующем в системе пользователе.
        Необязательное поле 'version' — версия, которую видел клиент; при расхождении с текущей вернётся 409
      parameters:
      - description: ID изменяемого пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно измененный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Нет или неверные данные Telegram
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_dadqeds_todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - TelegramInitData: []
      summary: Изменение пользователя
      tags:
      - users
securityDefinitions:
  TelegramInitData:
    description: '"tma <initData>" из Telegram.WebApp.initData. На локальном адресе
      (AUTH_LOCAL_ADDR) не требуется'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"